+++ Excel_PRIME.Tests/RowCellTests.cs
@@ -326,7 +326,11 @@ public class RowCellTests
                 IReadOnlyList<Cell>? rowCells = await row.GetAllCellsAsync().ConfigureAwait(true);
                 foreach (Cell cell in rowCells)
                 {
-                    if (cell.CellValue.IsUnknown) break;
+                    if (cell.CellValue.IsUnknown)
+                    {
+                        break;
+                    }
+
                     cell.CellValue.BoxedValue.Should().Be(workSheet1Content[r][c]);
                     c++;
                 }
@@ -373,7 +377,11 @@ public class RowCellTests
                 IReadOnlyList<Cell>? rowCells = await row.GetAllCellsAsync().ConfigureAwait(true);
                 foreach (Cell cell in rowCells)
                 {
-                    if (cell.CellValue.IsUnknown) break;
+                    if (cell.CellValue.IsUnknown)
+                    {
+                        break;
+                    }
+
                     cell.CellValue.BoxedValue.Should().Be(workSheet2Content[r][c]);
                     c++;
                 }