+++ Excel_PRIME/GlobalUsings.cs
@@ -1,6 +1,7 @@
 global using System;
 global using System.Collections.Generic;
 global using System.Linq;
+global using System.Threading.Tasks;
 
 // Make internal types visible to the benchmark project
 [assembly: System.Runtime.CompilerServices.InternalsVisibleTo("ExcelPRIME.Bench")]