+++ Excel_PRIME/Implementation/XmlSheetReader.cs
@@ -11,11 +11,12 @@ internal sealed class XmlSheetReader : IOpenXmlSheetReaderAsync
     private readonly InstanceContext _instanceContext;
     private readonly XmlReader _reader;
     private bool _isDisposed;
-    private readonly int _startRow;
+    private int _startRow;
     private readonly string _rowRefAtom;
     private readonly ReaderAtoms _readerAtoms;
+    private readonly XmlNameTable _sharedNameTable;
 
-    public XmlSheetReader(NonClosingStream stream, InstanceContext instanceContext, XmlNameTable sharedNameTable, CancellationToken ct)
+    public XmlSheetReader(NonClosingStream stream, InstanceContext instanceContext, XmlNameTable sharedNameTable)
     {
         _instanceContext = instanceContext;
         _reader = XmlReader.Create(stream, new XmlReaderSettings
@@ -31,6 +32,65 @@ internal sealed class XmlSheetReader : IOpenXmlSheetReaderAsync
             ValidationFlags = System.Xml.Schema.XmlSchemaValidationFlags.None,
             Async = true // TBD
         });
+        _sharedNameTable = sharedNameTable;
+        // Atomize key names once for fast lookups later.
+        _rowRefAtom = _sharedNameTable.Add("row");
+        _readerAtoms = new ReaderAtoms(_reader);
+    }
+
+    internal async Task InitializeAsync(CancellationToken ct)
+    {
+        string worksheetRefAtom = _reader.NameTable.Add("worksheet");
+        // Step into the worksheet
+        while (await _reader.ReadAsync().ConfigureAwait(false) && !ct.IsCancellationRequested)
+        {
+            if (_reader.NodeType == XmlNodeType.Element
+                && ReferenceEquals(_reader.LocalName, worksheetRefAtom)
+               )
+            {
+                break;
+            }
+        }
+
+        string dimensionRefAtom = _reader.NameTable.Add("dimension");
+        string colsRefAtom = _reader.NameTable.Add("cols");
+        string sheetDataRefAtom = _reader.NameTable.Add("sheetData");
+
+        bool foundSheetData = false;
+        while (!ct.IsCancellationRequested
+               && !foundSheetData   // Do not read after finding sheetData
+               && await _reader.ReadAsync().ConfigureAwait(false)
+              )
+        {
+            if (_reader.NodeType != XmlNodeType.Element)
+            {
+                continue;
+            }
+
+            string readerLocalName = _reader.LocalName;
+
+            if (ReferenceEquals(readerLocalName, dimensionRefAtom))
+            {
+                ParseDimension();
+            }
+            else if (ReferenceEquals(readerLocalName, colsRefAtom))
+            {
+                if (_reader.IsEmptyElement)
+                {
+                    // TODO: Need to understand when and how this is used
+                    //continue;
+                }
+            }
+            else if (ReferenceEquals(readerLocalName, sheetDataRefAtom))
+            {
+                foundSheetData = true;
+            }
+        }
+        CurrentRow = 0;
+    }
+
+    internal void Initialize(CancellationToken ct)
+    {
         string worksheetRefAtom = _reader.NameTable.Add("worksheet");
         // Step into the worksheet
         while (_reader.Read() && !ct.IsCancellationRequested)
@@ -62,31 +122,7 @@ internal sealed class XmlSheetReader : IOpenXmlSheetReaderAsync
 
             if (ReferenceEquals(readerLocalName, dimensionRefAtom))
             {
-                string? dim = _reader.GetAttribute("ref");
-                if (dim != null)
-                {
-                    ReadOnlySpan<char> dimSpan = dim.AsSpan();
-                    int colonIndex = dimSpan.IndexOf(':');
-                    if (colonIndex == -1)
-                    {
-                        (int rowExcel, int _, _) = dimSpan.GetRowColNumbers();
-                        _startRow = rowExcel - 1; // Take it back to the array offset
-                        SheetDimensions = (1, 1);
-                    }
-                    else
-                    {
-                        ReadOnlySpan<char> firstPart = dimSpan[..colonIndex];
-                        ReadOnlySpan<char> secondPart = dimSpan[(colonIndex + 1)..];
-                        (int rowExcel, int _, _) = firstPart.GetRowColNumbers();
-                        _startRow = rowExcel - 1; // Take it back to the array offset
-                        (int rowMax, int colMax, _) = secondPart.GetRowColNumbers();
-                        SheetDimensions = (rowMax, colMax);
-                    }
-                }
-                else
-                {
-                    SheetDimensions = new ValueTuple<int, int>(0, 0);
-                }
+                ParseDimension();
             }
             else if (ReferenceEquals(readerLocalName, colsRefAtom))
             {
@@ -102,9 +138,35 @@ internal sealed class XmlSheetReader : IOpenXmlSheetReaderAsync
             }
         }
         CurrentRow = 0;
-        // Atomize key names once for fast lookups later.
-        _rowRefAtom = sharedNameTable.Add("row");
-        _readerAtoms = new ReaderAtoms(_reader);
+    }
+
+    private void ParseDimension()
+    {
+        string? dim = _reader.GetAttribute("ref");
+        if (dim != null)
+        {
+            ReadOnlySpan<char> dimSpan = dim.AsSpan();
+            int colonIndex = dimSpan.IndexOf(':');
+            if (colonIndex == -1)
+            {
+                (int rowExcel, int _, _) = dimSpan.GetRowColNumbers();
+                _startRow = rowExcel - 1; // Take it back to the array offset
+                SheetDimensions = (1, 1);
+            }
+            else
+            {
+                ReadOnlySpan<char> firstPart = dimSpan[..colonIndex];
+                ReadOnlySpan<char> secondPart = dimSpan[(colonIndex + 1)..];
+                (int rowExcel, int _, _) = firstPart.GetRowColNumbers();
+                _startRow = rowExcel - 1; // Take it back to the array offset
+                (int rowMax, int colMax, _) = secondPart.GetRowColNumbers();
+                SheetDimensions = (rowMax, colMax);
+            }
+        }
+        else
+        {
+            SheetDimensions = (0, 0);
+        }
     }
 
     private static Row CreateRowFromPool()
@@ -123,8 +185,21 @@ internal sealed class XmlSheetReader : IOpenXmlSheetReaderAsync
                 return true;
             }
         }
-        if (_reader.EOF)
-        {   // No rows to read, or the Dimension is lying
+        return false;
+    }
+
+    private async Task<bool> ReadToNextStartRowAsync(CancellationToken ct)
+    {
+        while (await _reader.ReadToFollowingAsync(_rowRefAtom).ConfigureAwait(false)
+               && !ct.IsCancellationRequested
+              )
+        {
+            if (_reader.NodeType == XmlNodeType.Element
+                && ReferenceEquals(_reader.LocalName, _rowRefAtom)
+               )
+            {
+                return true;
+            }
         }
         return false;
     }
@@ -144,7 +219,7 @@ internal sealed class XmlSheetReader : IOpenXmlSheetReaderAsync
         }
     }
 
-    public (int Height, int Width) SheetDimensions { get; }
+    public (int Height, int Width) SheetDimensions { get; private set; }
 
     /// <summary>
     /// The Current row iterator offset (Starts at 1)
@@ -178,7 +253,7 @@ internal sealed class XmlSheetReader : IOpenXmlSheetReaderAsync
         }
 
         if (CurrentRow < _startRow
-            || !ReadToNextStartRow(ct)
+            || !await ReadToNextStartRowAsync(ct).ConfigureAwait(false)
            )
         {
             return null;